requirements.txt
attrs==23.2.0
bcrypt==4.1.3
click==8.1.7
Flask==2.2.5
Flask-JWT-Extended==4.6.0
greenlet==3.0.3
idna==3.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
isort==5.11.5
itsdangerous==2.1.2
Jinja2==3.1.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1 
MarkupSafe==2.1.5
multidict==6.0.5
mysqlclient==2.1.1
pkgutil_resolve_name==1.3.10
PyJWT==2.8.0
pyrsistent==0.19.3
python-dotenv==0.21.1
referencing==0.32.0
SQLAlchemy==2.0.31
typing_extensions==4.7.1
Werkzeug==2.2.3
yarl==1.9.4
zipp==3.15.0
